Move all functions to netlify folder
